feat: make dynamic-icon into a server component to reduce bundle size
